renders lists of links to content #30
