+++ web/controllers/article_controller.ex
@@ -3,30 +3,17 @@ defmodule App.ArticleController do
   import Ecto.Query, only: [from: 2]
   alias App.CMSRepo
 
-  def index(conn, _params) do
+  def show(conn, %{"id" => id}) do
     article_query = from a in "articles_articlepage",
-      select: a.body
+      where: a.page_ptr_id == ^String.to_integer(id),
+      select: %{heading: a.heading, body: a.body}
 
-    articles = CMSRepo.all(article_query)
-
-    render conn, "index.html", articles: articles
-  end
-
-  def show(conn, %{"tag" => tag}) do
-    article_query = from t in "taggit_tag",
-      where: t.name == ^tag,
-      join: apt in "articles_articlepagetag",
-      where: apt.tag_id == t.id,
-      join: a in "articles_articlepage",
-      where: a.page_ptr_id == apt.content_object_id,
-      select: a.body
-
-    case articles = CMSRepo.all(article_query) do
-      [] ->
+    case article = CMSRepo.one(article_query) do
+      nil ->
         conn
           |> put_status(404)
           |> render(App.ErrorView, "404.html")
-      _ -> render conn, "show.html", tag: tag, articles: articles
+      _ -> render conn, "show.html", article: article
     end
   end
 end
