--- web/templates/article/index.html.eex
@@ -1,5 +1,0 @@
-<h2>Listing Articles</h2>
-
-<%= for article <- @articles do %>
-  <%= raw(article) %>
-<% end %>
