+++ web/templates/article/show.html.eex
@@ -1,5 +1,3 @@
-<h2><%= String.capitalize @tag %> Articles</h2>
+<h2><%= assigns.article[:heading] %></h2>
 
-<%= for article <- @articles do %>
-  <%= raw(article) %>
-<% end %>
+<%= raw(assigns.article[:body]) %>
