+++ web/templates/article_list/index.html.eex
@@ -1,0 +1,7 @@
+<h2>All Articles</h2>
+
+<%= for article <- @articles do %>
+  <div>
+    <%= link article.heading, to: article_path(@conn, :show, article.id) %>
+  </div>
+<% end %>
