+++ web/views/article_list_view.ex
@@ -1,0 +1,3 @@
+defmodule App.ArticleListView do
+  use App.Web, :view
+end
